+++ modules/post/components/detail/participant/ParticipantCard.spec.tsx
@@ -3,11 +3,11 @@ import { render, screen } from '@testing-library/react-native';
 
 import mockMembers from '@post/mocks/members';
 import { wrapper } from '@shared/components/ThemeProvider';
-import Participants from './participants';
+import ParticipantCard from './ParticipantCard';
 
-describe('Participants Compoent', () => {
+describe('ParticipantCard Compoent', () => {
   test('멤버와 관련된 정보가 올바르게 렌더링 되어야 한다.', () => {
-    render(<Participants users={mockMembers} />, wrapper);
+    render(<ParticipantCard users={mockMembers} />, wrapper);
 
     expect(screen.getByText('모임 멤버')).toBeOnTheScreen();
