+++ modules/post/components/detail/participant/ParticipantCard.tsx
