+++ modules/post/components/detail/participant/member/style.ts
