+++ modules/post/components/detail/participant/type.tsx
